+++ src/database/models/Reservation.ts
@@ -1,11 +1,9 @@
-import { connection } from '../../src/connectDB';
+import { connection } from '../../connectDB';
 import { DataTypes } from 'sequelize';
 
 // npx sequelize-cli-typescript model:generate --name Reservation --attributes fk_id_room:integer,fk_id_user:integer,fk_id_schedule:integer,date:string
 
-
 const Reservation = connection.define('reservation', {
-
     id: {
         type: DataTypes.INTEGER,
         autoIncrement: true,
